+++ stream/stream.go
@@ -110,9 +110,17 @@ type Stream[T any] interface {
 	ToSlice() []T
 
 	// Iter makes the Stream compatible with Go's "for ... range" syntax.
-	// It's a function that must not be directly invoked but referenced in for ... range loops:
+	// This function must not be directly invoked but referenced in for ... range loops:
 	// for i, item := range Of(1, 2, 3).Iter { ... }
 	Iter(yield func(int, T) bool)
+
+	// Seq makes the Stream[T] compatible with the Go standard iter.Seq[T] type,
+	// allow using the Stream as an iterator that yields each element of this stream.
+	// It fulfills the standard iter.Seq[T] type definition and can be used with
+	// Go's "for ... range" syntax: for item := range stream.Seq { ... }
+	// as well as other functions using the standard Go iter.Seq type.
+	// This function must not be directly invoked but used as a reference.
+	Seq(yeld func(T) bool)
 }
 
 // if there are more items to iterate, returns the next item and true.